\name{carfima-internal}
\alias{g1}
\alias{g2}
\alias{B.mat}
\alias{V.mat.sigma}
\alias{V.mat}
\alias{Gamma.Y.sigma}
\alias{Gamma.Y}
\alias{carfima.loglik.nosigma}
\alias{carfima.bayes}

\title{Internal \code{carfima} functions}
\description{Internal \code{carfima} functions}
\details{These are not to be called by users.}
\keyword{internal}

